Move flipr base entity to its own file
